#221: add Raise a Toast to user context menus - improve accuracy of typing in UserContextMenuWrapper and MessageContextMenuWraper - fix import typo: PermissionFlagBits -> PermissionFlagsBits
