+++ ChangeLog.md
@@ -1,11 +1,15 @@
 # BountyBot Change Log
 ## BountyBot Version 2.8.0:
+### Context Menu Options
+- Added the following BountyBot functionality to the Apps dropdown when right-clicking on a User or Message:
+   - Check BountyBot Stats
+   - Raise a Toast
 ### End of Season Shoutouts
 When a season ends, the following bounty hunters will be recognized in the end of season message:
 - The hunter placed \#1 in season xp
 - The hunter who posted the most bounties completed in the season
 - The hunter who raised the most toasts in the season
-## New Items
+### New Items
 - Bonus Bounty Showcase: allows the hunter to showcase and increase the reward on one of their bounties on a separate cooldown from the command
 - Bounty Thumbnail: allows a bounty's poster to add a thumbnail image to one of their bounties (also increasing completion XP for the poster)
 ### Other Changes
